Rename chpl_arch -> chpl_cpu
