+++ util/chplenv/chpl_bin_subdir.py
@@ -4,7 +4,7 @@ import optparse
 import sys
 
 import chpl_home_utils
-import chpl_platform, chpl_machine, chpl_arch
+import chpl_platform, chpl_machine, chpl_cpu
 
 from utils import memoize
 
@@ -12,8 +12,8 @@ from utils import memoize
 def get(flag='host'):
     platform = chpl_platform.get(flag)
     machine = chpl_machine.get(flag)
-    (_, arch) = chpl_arch.get(flag, map_to_compiler=True,
-                              get_lcd=chpl_home_utils.using_chapel_module())
+    (_, arch) = chpl_cpu.get(flag, map_to_compiler=True,
+                             get_lcd=chpl_home_utils.using_chapel_module())
 
     # platform
     result = platform
