+++ util/chplenv/chplenv.py
@@ -1,4 +1,4 @@
-import chpl_arch
+import chpl_cpu
 import chpl_atomics
 import chpl_aux_filesys
 import chpl_bin_subdir
