+++ util/chplenv/third_party_utils.py
@@ -1,7 +1,7 @@
 import os
 import re
 
-import chpl_arch, chpl_machine, chpl_compiler
+import chpl_cpu, chpl_machine, chpl_compiler
 import chpl_lib_pic, chpl_locale_model, chpl_platform
 from chpl_home_utils import get_chpl_home, get_chpl_third_party, using_chapel_module
 from utils import error, memoize, run_command
@@ -12,8 +12,8 @@ from utils import error, memoize, run_command
 #
 @memoize
 def default_uniq_cfg_path():
-    arch_val = chpl_arch.get('target', map_to_compiler=True,
-                             get_lcd=using_chapel_module()).arch
+    arch_val = chpl_cpu.get('target', map_to_compiler=True,
+                            get_lcd=using_chapel_module()).arch
     return '{0}-{1}-{2}-{3}-{4}'.format(chpl_platform.get('target'),
                                         chpl_machine.get('target'),
                                         arch_val,
